pyproject.toml
[tool.poetry]
name = "magic_assistant"
version = "0.0.1"
description = "An AI agent framework."
authors = ["luguanglong <luguang_long@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
llama-cpp-python = "^0.1.68"
case-converter = "^1.1.0"
sqlalchemy = "^2.0.19"
pgvector = "^0.2.0"
loguru = "^0.7.0"
transformers = "^4.31.0"
sqlalchemy-utils = "^0.41.1"
torch = "^2.0.1"
nvidia-cublas-cu11 = "11.10.3.66"
sentence-transformers = "^2.2.2"
rich = "^13.5.2"
accelerate = "^0.21.0"
psycopg2 = "^2.9.7"
websockets = "^11.0.3"
mkdocs = "^1.5.2"
mkdocs-material = "^9.2.3"
python-docx = "^0.8.11"
python-pptx = "^0.6.21"
fitz = "^0.0.1.dev2"
minio = "^7.1.16"
pymupdf = "^1.23.1"
python-multipart = "^0.0.6"
chardet = "^5.2.0"
pymilvus = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#设置poetry包管理工具的自定义pypi镜像源配置
[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
